:fire: fix page routes and feedback changes
