+++ components/nav.js
@@ -1,12 +1,14 @@
 import Link from "next/link";
 import styles from "../styles/navstyle.module.scss";
 import { navLinks } from "../constants/nav_links";
-import { HELP_CENTER } from "../constants/page_links";
+import { useRouter } from "next/router";
 export default function Nav() {
+  const router = useRouter();
+  const pageUrl = router.pathname;
   return (
     <nav className={styles.navWrapper}>
       <div className={styles.brandLogo}>
-        <img src="/assets/images/EDD.png" />
+        <img src='/assets/images/EDD.png' />
       </div>
       <ul>
         {navLinks.map((_o, idx) => {
@@ -16,7 +18,7 @@ export default function Nav() {
               <Link href={path}>
                 <a
                   className={
-                    enabled && path == HELP_CENTER ? styles.active_link : ""
+                    enabled && path == pageUrl ? styles.active_link : ""
                   }
                 >
                   {text}
