+++ constants/page_links.js
@@ -3,7 +3,7 @@ export const CLAIMS = "";
 export const EMPLOYERS = "";
 export const NEWS_ROOM = "";
 export const HELP_CENTER = "";
-export const CONTACT = "";
+export const CONTACT = "/contact-us";
 export const ONLINE_SERVICES = "";
 export const LABOR_MARKET_INFORMATION = "";
 export const FORMS_PUBLICATIONS = "";
